Trampoline Can be Used in the Air
The trampoline can be used on the ground or in the air now, but once the
player jumps on it, he cannot use it again until he touches the ground.
Also, fixed the bug where the player can still jump if he runs off a
cliff without jumping or using a trampoline.